+++ android/src/main/kotlin/com/chizi/carrier_info/MethodCallHandlerImpl.kt
@@ -8,23 +8,18 @@ import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Handler
 import android.os.Looper
-import android.telephony.TelephonyManager
+import android.telephony.*
+import androidx.annotation.RequiresApi
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import io.flutter.Log
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-/*import android.telephony.CellInfoLte
-import android.telephony.CellInfoCdma
-import android.telephony.CellInfoGsm*/
-import android.telephony.cdma.CdmaCellLocation 
-import android.telephony.gsm.GsmCellLocation 
 
 
 internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : MethodCallHandler {
-
-
-    private val TAG: String =  "carrier_info"
+    private val TAG: String = "carrier_info"
     private var context: Context?
     private var activity: Activity?
     private val E_NO_CARRIER_NAME = "no_carrier_name"
@@ -34,9 +29,8 @@ internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : Me
     private val E_NO_MOBILE_NETWORK = "no_mobile_network"
     private val E_NO_NETWORK_OPERATOR = "no_network_operator"
     private val E_NO_CELL_ID = "no_cell_id"
-    private var mTelephonyManager: TelephonyManager? = null
+    private var mDefaultTelephonyManager: TelephonyManager? = null
     private lateinit var func: () -> Unit?
-
 
 
     fun setActivity(act: Activity?) {
@@ -47,59 +41,22 @@ internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : Me
         this.activity = activity
         this.context = context
 
-        mTelephonyManager = context.getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
-
+        // Retrieve the default subscription's TelephonyManager used for all calls"
+        mDefaultTelephonyManager =
+            context.getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
     }
 
     override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
         Handler(Looper.getMainLooper()).post {
             when (call.method) {
-                "carrierName" -> {
-                    carrierName(result)
-                }
-                "isoCountryCode" -> {
-                    isoCountryCode(result)
-                }
-                "mobileCountryCode" -> {
-                    mobileCountryCode(result)
-                }
-                "mobileNetworkCode" -> {
-                    mobileNetworkCode(result)
-                }
-                "mobileNetworkOperator" -> {
-                    mobileNetworkOperator(result)
-                }
-                "radioType" -> {
-                    func = { radioType(result) }
-                    if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
-                        if (!checkIfAlreadyHavePermission()) {
-                            requestForSpecificPermission(0)
-                        }else{
-                            radioType(result)
-                        }
+                "getAndroidInfo" -> {
+                    try {
+                        getInfo(result)
+                    } catch (e: Exception) {
+                        result.error(E_NO_CARRIER_NAME, "No carrier name", e.toString())
                     }
                 }
-                "networkGeneration" -> {
-                    func = { networkGeneration(result) }
-                    if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
-                        if (!checkIfAlreadyHavePermission()) {
-                            requestForSpecificPermission(1)
-                        }else{
-                            networkGeneration(result)
-                        }
-                    }
 
-                }
-                "cellId" -> {
-                    func = { cellId(result) }
-                    if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
-                        if (!checkIfAlreadyHavePermission()) {
-                            requestForSpecificPermission(2)
-                        }else{
-                            cellId(result)
-                        }
-                    }
-                }
                 else -> {
                     result.notImplemented()
                 }
@@ -107,182 +64,277 @@ internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : Me
         }
     }
 
+
     private fun requestForSpecificPermission(i: Int) {
-        ActivityCompat.requestPermissions(this.activity!!, arrayOf(Manifest.permission.READ_PHONE_STATE, Manifest.permission.ACCESS_NETWORK_STATE, Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION), i)
+        ActivityCompat.requestPermissions(
+            this.activity!!,
+            arrayOf(
+                Manifest.permission.READ_PHONE_STATE,
+                Manifest.permission.ACCESS_NETWORK_STATE,
+                Manifest.permission.ACCESS_COARSE_LOCATION,
+                Manifest.permission.ACCESS_FINE_LOCATION
+            ),
+            i
+        )
     }
 
     private fun checkIfAlreadyHavePermission(): Boolean {
-        return ContextCompat.checkSelfPermission(this.activity!!, Manifest.permission.READ_PHONE_STATE) == PackageManager.PERMISSION_GRANTED && 
-        ContextCompat.checkSelfPermission(this.activity!!, Manifest.permission.ACCESS_NETWORK_STATE) == PackageManager.PERMISSION_GRANTED && 
-        ContextCompat.checkSelfPermission(this.activity!!, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED &&
-        ContextCompat.checkSelfPermission(this.activity!!, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED
+        return ContextCompat.checkSelfPermission(
+            this.activity!!,
+            Manifest.permission.READ_PHONE_STATE
+        ) == PackageManager.PERMISSION_GRANTED &&
+                ContextCompat.checkSelfPermission(
+                    this.activity!!,
+                    Manifest.permission.ACCESS_NETWORK_STATE
+                ) == PackageManager.PERMISSION_GRANTED &&
+                ContextCompat.checkSelfPermission(
+                    this.activity!!,
+                    Manifest.permission.ACCESS_COARSE_LOCATION
+                ) == PackageManager.PERMISSION_GRANTED &&
+                ContextCompat.checkSelfPermission(
+                    this.activity!!,
+                    Manifest.permission.ACCESS_FINE_LOCATION
+                ) == PackageManager.PERMISSION_GRANTED
     }
 
-    private fun carrierName(result: MethodChannel.Result) {
-        val carrierName = mTelephonyManager!!.simOperatorName
-        if (carrierName != null && "" != carrierName) {
-            result.success(carrierName)
-        } else {
-            result.error(E_NO_CARRIER_NAME, "No carrier name","")
-        }
-    }
 
-    private fun radioType(result: MethodChannel.Result) {
-        when (mTelephonyManager!!.getDataNetworkType()) {
-            TelephonyManager.NETWORK_TYPE_1xRTT -> result.success("1xRTT")
-            TelephonyManager.NETWORK_TYPE_CDMA -> result.success("CDMA")
-            TelephonyManager.NETWORK_TYPE_EDGE -> result.success("EDGE")
-            TelephonyManager.NETWORK_TYPE_EHRPD -> result.success("eHRPD")
-            TelephonyManager.NETWORK_TYPE_EVDO_0 ->result.success("EVDO rev. 0")
-            TelephonyManager.NETWORK_TYPE_EVDO_A -> result.success("EVDO rev. A")
-            TelephonyManager.NETWORK_TYPE_EVDO_B -> result.success("EVDO rev. B")
-            TelephonyManager.NETWORK_TYPE_GPRS -> result.success("GPRS")
-            TelephonyManager.NETWORK_TYPE_GSM -> result.success("GSM")
-            TelephonyManager.NETWORK_TYPE_HSDPA -> result.success("HSDPA")
-            TelephonyManager.NETWORK_TYPE_HSPA -> result.success("HSPA")
-            TelephonyManager.NETWORK_TYPE_HSPAP -> result.success("HSPA+")
-            TelephonyManager.NETWORK_TYPE_HSUPA -> result.success("HSUPA")
-            TelephonyManager.NETWORK_TYPE_IDEN -> result.success("iDen")
-            TelephonyManager.NETWORK_TYPE_UMTS -> result.success("UMTS")
-            TelephonyManager.NETWORK_TYPE_LTE -> result.success("LTE")
-            TelephonyManager.NETWORK_TYPE_NR -> result.success("NR")
-            TelephonyManager.NETWORK_TYPE_TD_SCDMA -> result.success("TD SCDMA")
-            TelephonyManager.NETWORK_TYPE_IWLAN -> result.success("IWLAN")
-            TelephonyManager.NETWORK_TYPE_UNKNOWN -> result.success("Unknown")
-        }
-    }
+    private fun getInfo(result: MethodChannel.Result) {
 
-    private fun networkGeneration(result: MethodChannel.Result) {
-        val radioType = mTelephonyManager?.getDataNetworkType()
-        if (radioType != null) {
-            when (radioType) {
-                TelephonyManager.NETWORK_TYPE_GPRS,
-                TelephonyManager.NETWORK_TYPE_EDGE,
-                TelephonyManager.NETWORK_TYPE_CDMA,
-                TelephonyManager.NETWORK_TYPE_1xRTT,
-                TelephonyManager.NETWORK_TYPE_IDEN,
-                TelephonyManager.NETWORK_TYPE_GSM
-                -> result.success("2G")
-                TelephonyManager.NETWORK_TYPE_UMTS,
-                TelephonyManager.NETWORK_TYPE_EVDO_0,
-                TelephonyManager.NETWORK_TYPE_EVDO_A,
-                TelephonyManager.NETWORK_TYPE_HSDPA,
-                TelephonyManager.NETWORK_TYPE_HSUPA,
-                TelephonyManager.NETWORK_TYPE_HSPA,
-                TelephonyManager.NETWORK_TYPE_EVDO_B,
-                TelephonyManager.NETWORK_TYPE_EHRPD,
-                TelephonyManager.NETWORK_TYPE_HSPAP,
-                TelephonyManager.NETWORK_TYPE_TD_SCDMA
-                -> result.success("3G")
-                TelephonyManager.NETWORK_TYPE_LTE
-                -> result.success("4G")
-                TelephonyManager.NETWORK_TYPE_NR
-                -> result.success("5G")
-                else -> result.success("Unknown")
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            val telephonyList: ArrayList<HashMap<String, Any?>> = ArrayList()
+            val subscriptionsList: ArrayList<HashMap<String, Any?>> = ArrayList()
+            val subsManager =
+                context!!.getSystemService(Context.TELEPHONY_SUBSCRIPTION_SERVICE) as SubscriptionManager
+
+            if (subsManager.activeSubscriptionInfoList != null) {
+
+                for (subsInfo in subsManager.activeSubscriptionInfoList) {
+                    if (subsInfo != null) {
+                        try {
+                            val data = hashMapOf(
+                                "simSerialNo" to subsInfo.iccId,
+                                "mobileCountryCode" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.mccString else subsInfo.mcc.toString(),
+                                "countryIso" to subsInfo.countryIso,
+                                "mobileNetworkCode" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.mncString else subsInfo.mnc.toString(),
+                                "cardId" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.cardId else null,
+                                "carrierId" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.carrierId else null,
+                                "dataRoaming" to subsInfo.dataRoaming,
+                                "isEmbedded" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.isEmbedded else null,
+                                "simSlotIndex" to subsInfo.simSlotIndex,
+                                "subscriptionType" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.subscriptionType else null,
+                                "isOpportunistic" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) subsInfo.isOpportunistic else null,
+                                "displayName" to subsInfo.displayName,
+                                "subscriptionId" to subsInfo.subscriptionId,
+                                "phoneNumber" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)
+                                    subsManager.getPhoneNumber(subsInfo.subscriptionId) else subsInfo.number,
+                                "isNetworkRoaming" to subsManager.isNetworkRoaming(subsInfo.subscriptionId),
+                            )
+
+                            subscriptionsList.add(data)
+                        } catch (e: Exception) {
+                            Log.d(
+                                TAG,
+                                "SubscriptionInfo Exception: ${subsInfo.carrierName}\n${e}"
+                            )
+                            subscriptionsList.add(HashMap())
+                        }
+
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                            val telephonyManager =
+                                mDefaultTelephonyManager!!.createForSubscriptionId(subsInfo.subscriptionId);
+
+                            try {
+                                val data = hashMapOf<String, Any?>(
+                                    "carrierName" to telephonyManager.simOperatorName,
+                                    "dataActivity" to telephonyManager.dataActivity,
+                                    "radioType" to radioType(telephonyManager),
+                                    "cellId" to cellId(telephonyManager, subsInfo.simSlotIndex),
+                                    "simState" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) simState(
+                                        telephonyManager
+                                    ) else null,
+                                    "phoneNumber" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)
+                                        subsManager.getPhoneNumber(subsInfo.subscriptionId) else subsInfo.number,
+                                    "networkOperatorName" to telephonyManager.networkOperatorName,
+                                    "subscriptionId" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) telephonyManager.subscriptionId else null,
+                                    "isoCountryCode" to telephonyManager.simCountryIso,
+                                    "networkCountryIso" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) telephonyManager.networkCountryIso else null,
+                                    "mobileNetworkCode" to if (telephonyManager.simOperator?.isEmpty() == false)
+                                        telephonyManager.simOperator?.substring(3) else null,
+                                    "displayName" to telephonyManager.simOperatorName,
+                                    "mobileCountryCode" to if (telephonyManager.simOperator?.isEmpty() == false)
+                                        telephonyManager.simOperator?.substring(0, 3) else null,
+                                    "networkGeneration" to networkGeneration(telephonyManager),
+                                )
+
+                                telephonyList.add(data)
+                            } catch (e: Exception) {
+                                Log.d(
+                                    TAG,
+                                    "TelephonyManager Exception: ${telephonyManager.simOperatorName}\n${e}"
+                                )
+                                telephonyList.add(HashMap())
+                            }
+                        }
+                    }
                 }
+            }
 
+            val data = hashMapOf<String, Any?>(
+                "isDataEnabled" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) mDefaultTelephonyManager!!.isDataEnabled else null,
+                "isMultiSimSupported" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) isMultiSimSupported() else null,
+                "isDataCapable" to if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) mDefaultTelephonyManager!!.isDataCapable else null,
+                "isSmsCapable" to mDefaultTelephonyManager!!.isSmsCapable,
+                "isVoiceCapable" to mDefaultTelephonyManager!!.isVoiceCapable,
+                "telephonyInfo" to telephonyList,
+                "subscriptionsInfo" to subscriptionsList,
+            )
+
+            result.success(data)
         } else {
-            result.error(E_NO_NETWORK_TYPE, "No network type","")
+            result.error(E_NO_CARRIER_NAME, "No carrier name", "")
+        }
+
+
+    }
+
+
+    @RequiresApi(Build.VERSION_CODES.N)
+    private fun radioType(telephonyManager: TelephonyManager): String {
+        return when (telephonyManager.dataNetworkType) {
+            TelephonyManager.NETWORK_TYPE_1xRTT -> return "1xRTT"
+            TelephonyManager.NETWORK_TYPE_CDMA -> return "CDMA"
+            TelephonyManager.NETWORK_TYPE_EDGE -> return "EDGE"
+            TelephonyManager.NETWORK_TYPE_EHRPD -> return "eHRPD"
+            TelephonyManager.NETWORK_TYPE_EVDO_0 -> return "EVDO rev. 0"
+            TelephonyManager.NETWORK_TYPE_EVDO_A -> return "EVDO rev. A"
+            TelephonyManager.NETWORK_TYPE_EVDO_B -> return "EVDO rev. B"
+            TelephonyManager.NETWORK_TYPE_GPRS -> return "GPRS"
+            TelephonyManager.NETWORK_TYPE_GSM -> return "GSM"
+            TelephonyManager.NETWORK_TYPE_HSDPA -> return "HSDPA"
+            TelephonyManager.NETWORK_TYPE_HSPA -> return "HSPA"
+            TelephonyManager.NETWORK_TYPE_HSPAP -> return "HSPA+"
+            TelephonyManager.NETWORK_TYPE_HSUPA -> return "HSUPA"
+            TelephonyManager.NETWORK_TYPE_IDEN -> return "iDen"
+            TelephonyManager.NETWORK_TYPE_UMTS -> return "UMTS"
+            TelephonyManager.NETWORK_TYPE_LTE -> return "LTE"
+            TelephonyManager.NETWORK_TYPE_NR -> return "NR"
+            TelephonyManager.NETWORK_TYPE_TD_SCDMA -> return "TD SCDMA"
+            TelephonyManager.NETWORK_TYPE_IWLAN -> return "IWLAN"
+            TelephonyManager.NETWORK_TYPE_UNKNOWN -> return "Unknown"
+            else -> ""
         }
     }
 
-    private fun isoCountryCode(result: MethodChannel.Result) {
-        val iso = mTelephonyManager!!.simCountryIso
-        if (iso != null && "" != iso) {
-            result.success(iso)
-        } else {
-            result.error(E_NO_ISO_COUNTRY_CODE, "No iso country code","")
+
+    @RequiresApi(Build.VERSION_CODES.Q)
+    private fun isMultiSimSupported(): String {
+        return when (mDefaultTelephonyManager!!.isMultiSimSupported) {
+            TelephonyManager.MULTISIM_ALLOWED -> return "MULTISIM_ALLOWED"
+            TelephonyManager.MULTISIM_NOT_SUPPORTED_BY_CARRIER -> return "MULTISIM_NOT_SUPPORTED_BY_CARRIER"
+            TelephonyManager.MULTISIM_NOT_SUPPORTED_BY_HARDWARE -> return "MULTISIM_NOT_SUPPORTED_BY_HARDWARE"
+            else -> ""
         }
     }
 
-    // returns MCC (3 digits)
 
-    private fun mobileCountryCode(result: MethodChannel.Result) {
-        val plmn = mTelephonyManager!!.simOperator
-        if (plmn != null && "" != plmn) {
-            result.success(plmn.substring(0, 3))
-        } else {
-            result.error(E_NO_MOBILE_COUNTRY_CODE, "No mobile country code","")
+    @RequiresApi(Build.VERSION_CODES.O)
+    private fun simState(telephonyManager: TelephonyManager): String {
+        return when (telephonyManager.simState) {
+            TelephonyManager.SIM_STATE_ABSENT -> return "SIM_STATE_ABSENT"
+            TelephonyManager.SIM_STATE_CARD_IO_ERROR -> return "SIM_STATE_CARD_IO_ERROR"
+            TelephonyManager.SIM_STATE_CARD_RESTRICTED -> return "SIM_STATE_CARD_RESTRICTED"
+            TelephonyManager.SIM_STATE_NETWORK_LOCKED -> return "SIM_STATE_NETWORK_LOCKED"
+            TelephonyManager.SIM_STATE_NOT_READY -> return "SIM_STATE_NOT_READY"
+            TelephonyManager.SIM_STATE_PERM_DISABLED -> return "SIM_STATE_PERM_DISABLED"
+            TelephonyManager.SIM_STATE_PIN_REQUIRED -> return "SIM_STATE_PIN_REQUIRED"
+            TelephonyManager.SIM_STATE_PUK_REQUIRED -> return "SIM_STATE_PUK_REQUIRED"
+            TelephonyManager.SIM_STATE_READY -> return "SIM_STATE_READY"
+            TelephonyManager.SIM_STATE_UNKNOWN -> return "SIM_STATE_UNKNOWN"
+            else -> ""
         }
     }
 
-    // returns MNC (2 or 3 digits)
+    @RequiresApi(Build.VERSION_CODES.N)
+    private fun networkGeneration(telephonyManager: TelephonyManager): String {
+        when (val radioType = telephonyManager.dataNetworkType) {
+            TelephonyManager.NETWORK_TYPE_GPRS,
+            TelephonyManager.NETWORK_TYPE_EDGE,
+            TelephonyManager.NETWORK_TYPE_CDMA,
+            TelephonyManager.NETWORK_TYPE_1xRTT,
+            TelephonyManager.NETWORK_TYPE_IDEN,
+            TelephonyManager.NETWORK_TYPE_GSM
+            -> return "2G"
 
-    private fun mobileNetworkCode(result: MethodChannel.Result) {
-        val plmn = mTelephonyManager!!.simOperator
-        if (plmn != null && "" != plmn) {
-            result.success(plmn.substring(3))
-        } else {
-            result.error(E_NO_MOBILE_NETWORK, "No mobile network code","")
-        }
-    }
+            TelephonyManager.NETWORK_TYPE_UMTS,
+            TelephonyManager.NETWORK_TYPE_EVDO_0,
+            TelephonyManager.NETWORK_TYPE_EVDO_A,
+            TelephonyManager.NETWORK_TYPE_HSDPA,
+            TelephonyManager.NETWORK_TYPE_HSUPA,
+            TelephonyManager.NETWORK_TYPE_HSPA,
+            TelephonyManager.NETWORK_TYPE_EVDO_B,
+            TelephonyManager.NETWORK_TYPE_EHRPD,
+            TelephonyManager.NETWORK_TYPE_HSPAP,
+            TelephonyManager.NETWORK_TYPE_TD_SCDMA
+            -> return "3G"
 
-    // return MCC + MNC (5 or 6 digits), e.g. 20601
-    private fun mobileNetworkOperator(result: MethodChannel.Result){
-        val plmn = mTelephonyManager!!.simOperator
-        if (plmn != null && "" != plmn) {
-            result.success(plmn)
-        } else {
-            result.error(E_NO_NETWORK_OPERATOR, "No mobile network operator", "")
+            TelephonyManager.NETWORK_TYPE_LTE
+            -> return "4G"
+
+            TelephonyManager.NETWORK_TYPE_NR,
+            -> return "5G"
+
+            else -> radioType.toString()
         }
+        return "unknown"
     }
 
     // return cell id
-    private fun cellId(result: MethodChannel.Result){
-        val location = mTelephonyManager!!.getCellLocation()
-        if (location != null) {
-            var cid = -1
-            var lac = -1
-            if (location is GsmCellLocation) {
-                cid = (location as GsmCellLocation).getCid()
-                lac = (location as GsmCellLocation).getLac()
+    private fun cellId(telephonyManager: TelephonyManager, slotIndex: Int): HashMap<String, Any>? {
 
-            } else if(location is CdmaCellLocation){
-                cid = (location as CdmaCellLocation).getBaseStationId()
-                lac = (location as CdmaCellLocation).getNetworkId()
-            }
-            result.success(
-                """
-                {
-                    "cid": $cid,
-                    "lac": $lac
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
+            // registeredCellInfo elements are ordered, simSlot 0 information will be in the index 0
+            val registeredCellInfo: ArrayList<CellInfo> = ArrayList()
+            for (cellInfo in telephonyManager.allCellInfo) {
+                if (cellInfo.isRegistered) {
+                    registeredCellInfo.add(cellInfo)
                 }
-                """
-            )
-            return
-        }
-        // FOLLOWING CODE WITH THE NEW API BUT DOES NOT WORK ON Android 11 AOSP //
-        /*
-        val cellInfos = mTelephonyManager!!.getAllCellInfo()
-        if (cellInfos != null && cellInfos.size > 0) {
-            val info = cellInfos[0]
-            var cid = -1
-            var lac = -1
-            if (info is CellInfoGsm) {
-                cid = (info as CellInfoGsm).getCellIdentity().getCid()
-                lac = (info as CellInfoGsm).getCellIdentity().getLac()
-            } else if(info is CellInfoCdma){
-                cid = (info as CellInfoCdma).getCellIdentity().getBasestationId()
-                lac = (info as CellInfoCdma).getCellIdentity().getNetworkId()
-            } else if(info is CellInfoLte) {
-                cid = (info as CellInfoLte).getCellIdentity().getCi()
             }
-            result.success(
-                """
-                {
-                    "cid": $cid,
-                    "lac": $lac
-                }
-                """
-            )
-            return
+
+            var cid = -1;
+            var lac = -1;
+
+            val currentCellInfo = registeredCellInfo[slotIndex]
+
+            if (currentCellInfo is CellInfoGsm) {
+                val identityGsm = currentCellInfo.cellIdentity
+                cid = identityGsm.cid
+                lac = identityGsm.lac
+            } else if (currentCellInfo is CellInfoCdma) {
+                val identityCdma = currentCellInfo.cellIdentity
+                cid = identityCdma.basestationId
+                lac = identityCdma.networkId
+            } else if (currentCellInfo is CellInfoLte) {
+                val identityLte = currentCellInfo.cellIdentity
+                cid = identityLte.ci
+                lac = identityLte.tac
+            } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2 && currentCellInfo is CellInfoWcdma) {
+                val identityWcdma = currentCellInfo.cellIdentity
+                cid = identityWcdma.cid
+                lac = identityWcdma.lac
+            }
+
+            return hashMapOf("cid" to cid, "lac" to lac)
         }
-        */
-        result.error(E_NO_CELL_ID, "No cell id", "")
+
+        return null;
     }
 
 
-    fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>?, grantResults: IntArray?) {
+    fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<out String>?,
+        grantResults: IntArray?
+    ) {
         when (requestCode) {
             0 -> return if (grantResults!![0] == PackageManager.PERMISSION_GRANTED) {
                 this.func()!!
@@ -291,6 +343,7 @@ internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : Me
                 requestForSpecificPermission(0)
 
             }
+
             1 -> return if (grantResults!![0] == PackageManager.PERMISSION_GRANTED) {
                 this.func()!!
 
@@ -298,6 +351,7 @@ internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : Me
                 requestForSpecificPermission(1)
 
             }
+
             2 -> return if (grantResults!![0] == PackageManager.PERMISSION_GRANTED) {
                 this.func()!!
 
@@ -307,3 +361,4 @@ internal class MethodCallHandlerImpl(context: Context, activity: Activity?) : Me
         }
     }
 }
+